Moved limit-service sources from lib to src
A mechanical move to the layout every internal package uses, kept on its own so the
conversion that follows reads as a change to the code rather than a change of address.
Only the references the move itself invalidates are touched, and the package behaves
exactly as it did.
