--- packages/limit-service/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/LimitService');
